refactor(app): fold the workspace helpers into one service
Three modules of loose functions — download tokens, skill uploads, git
status — became `WorkspaceService`, injected from `app.state` like
`SessionService`. The router keeps only HTTP concerns.

The methods take what a caller has, not what the implementation wants:

- `read_status(user_id, agent_id, session_id)` replaces a
  `read_git_status(backend, cwd)` that made every caller resolve the
  session, the workspace and the cwd itself. `WorkspaceStatus` moves
  next to `GitStatus`, since the service is what produces it.
- `resolve()` returns the workspace alone. Eleven of its twelve callers
  discarded the session record the old tuple carried.
- `install_slots` was a public semaphore both install paths had to
  remember to hold; forgetting it silently removed the limit. It is now
  private, held inside `install_skill()`, which takes a workspace
  because the from-library path resolves once and installs several.
- The manifest/parts count check moves out of the router into
  `validate_manifest()`, alongside every other upload check.

`_trim_for_listing` and `_resolve_workspace` are inlined; neither did
anything its one call site could not say directly.

Frontend files are `pnpm format` output only.
