+++ src/agentscope/app/_router/_schema/_workspace.py
@@ -3,7 +3,6 @@
 from pydantic import BaseModel, Field
 
 from ....mcp import MCPClient
-from ..._service import GitStatus
 
 
 class AddSkillRequest(BaseModel):
@@ -85,35 +84,6 @@ class DirectoryListing(BaseModel):
     )
     entries: list[DirectoryEntry] = Field(
         description="The directory's immediate children, unsorted.",
-    )
-
-
-class WorkspaceStatus(BaseModel):
-    """Where a session is pointed, and the git state of that place."""
-
-    workdir: str = Field(
-        description=(
-            "Absolute path of the workspace root. Backend-dependent — a "
-            "host directory locally, a fixed path inside the sandbox "
-            "otherwise — so the client cannot derive it."
-        ),
-    )
-    cwd: str = Field(
-        description=(
-            "Absolute path the session is focused on, resolved from "
-            ":attr:`SessionConfig.cwd`. Equals :attr:`workdir` when that "
-            "is unset."
-        ),
-    )
-    git: GitStatus | None = Field(
-        default=None,
-        description=(
-            "Git state of :attr:`cwd`, or null when there is none to "
-            "report — not a repository, git unavailable, the command "
-            "timed out, or the backend could not run it. The caller "
-            "shows no branch either way, so the reasons are logged "
-            "rather than returned."
-        ),
     )
