--- src/agentscope/app/_service/_download_token.py
@@ -1,126 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Short-lived capability tokens for browser-native file downloads.
-
-A browser only streams a response to disk when it issues the request
-itself — a navigation or ``<a download>`` — and such a request cannot
-carry a custom header. Fetching with ``X-User-ID`` and building a blob
-works, but buffers the whole file in the tab.
-
-So the credential has to travel in the URL. A token minted here says
-"this exact user was already authenticated, and may read this exact
-path for the next minute" — a capability, not an identity. Only
-``GET /workspace/files`` accepts one, and minting depends on the normal
-identity dependency, so swapping in OAuth/JWT later changes where the
-capability comes from without touching how it is checked.
-"""
-import base64
-import hashlib
-import hmac
-import time
-
-# Long enough to survive a slow round trip and the user's click, short
-# enough that a token leaked through an access log or browser history
-# is already dead. There is no revocation list to fall back on.
-DEFAULT_DOWNLOAD_TOKEN_TTL = 60
-
-
-def _b64(raw: bytes) -> str:
-    """Encode without padding, which is not URL-safe to round-trip."""
-    return base64.urlsafe_b64encode(raw).decode("ascii").rstrip("=")
-
-
-def _unb64(text: str) -> bytes:
-    """Reverse :func:`_b64`, restoring the stripped padding."""
-    return base64.urlsafe_b64decode(text + "=" * (-len(text) % 4))
-
-
-def _signature(secret: str, expires_at: int, user_id: str, path: str) -> bytes:
-    """Compute the MAC binding an expiry, a user and a path together.
-
-    ``\\0`` separates the fields because it cannot occur in any of them,
-    so no combination of values can be re-cut into a different triple.
-
-    Args:
-        secret (`str`):
-            The app-wide signing secret.
-        expires_at (`int`):
-            Unix timestamp after which the token is refused.
-        user_id (`str`):
-            The user the capability was granted to.
-        path (`str`):
-            The one path the capability covers.
-
-    Returns:
-        `bytes`:
-            The raw HMAC-SHA256 digest.
-    """
-    message = f"{expires_at}\0{user_id}\0{path}".encode("utf-8")
-    return hmac.new(secret.encode("utf-8"), message, hashlib.sha256).digest()
-
-
-def sign_download_token(
-    secret: str,
-    user_id: str,
-    path: str,
-    ttl: int = DEFAULT_DOWNLOAD_TOKEN_TTL,
-) -> tuple[str, int]:
-    """Mint a token authorizing one download.
-
-    Args:
-        secret (`str`):
-            The app-wide signing secret.
-        user_id (`str`):
-            The already-authenticated caller.
-        path (`str`):
-            The path the token authorizes, verbatim as the download
-            request will send it.
-        ttl (`int`, defaults to 60):
-            Seconds the token stays valid.
-
-    Returns:
-        `tuple[str, int]`:
-            The token and its expiry as a Unix timestamp.
-    """
-    expires_at = int(time.time()) + ttl
-    signature = _signature(secret, expires_at, user_id, path)
-    token = f"{expires_at}.{_b64(user_id.encode('utf-8'))}.{_b64(signature)}"
-    return token, expires_at
-
-
-def verify_download_token(secret: str, token: str, path: str) -> str:
-    """Return the user a token was granted to, for this path.
-
-    The path is not read out of the token but re-derived from the
-    request, so a token for one file cannot be replayed against
-    another.
-
-    Args:
-        secret (`str`):
-            The app-wide signing secret.
-        token (`str`):
-            The token from the request.
-        path (`str`):
-            The path the request is asking for.
-
-    Returns:
-        `str`:
-            The user ID the token was minted for.
-
-    Raises:
-        `ValueError`:
-            The token is malformed, expired, or does not match.
-    """
-    try:
-        raw_expiry, raw_user, raw_signature = token.split(".")
-        expires_at = int(raw_expiry)
-        user_id = _unb64(raw_user).decode("utf-8")
-        signature = _unb64(raw_signature)
-    except (ValueError, UnicodeDecodeError) as e:
-        raise ValueError("Malformed download token.") from e
-
-    expected = _signature(secret, expires_at, user_id, path)
-    if not hmac.compare_digest(signature, expected):
-        raise ValueError("Invalid download token.")
-    if expires_at < time.time():
-        raise ValueError("Expired download token.")
-    return user_id
